Formulaire infos club ok

+++ gestadh45.dao/AdherentDao.cs
@@ -178,7 +178,7 @@ namespace gestadh45.dao
 			cmd.Parameters.Add(paramPrenom);
 
 			try {
-				var result = (int)cmd.ExecuteScalar();
+				var result = (long)cmd.ExecuteScalar();
 				return result > 0;
 			}
 			catch (SQLiteException) {
@@ -197,7 +197,7 @@ namespace gestadh45.dao
 			cmd.Parameters.Add(paramId);
 
 			try {
-				var result = (int)cmd.ExecuteScalar();
+				var result = (long)cmd.ExecuteScalar();
 				return result > 0;
 			}
 			catch (SQLiteException) {
