+++ gestadh45.dao/InscriptionDao.cs
@@ -105,7 +105,7 @@ namespace gestadh45.dao
 			cmd.Parameters.Add(paramIdGroupe);
 
 			try {
-				var result = (int)cmd.ExecuteScalar();
+				var result = (long)cmd.ExecuteScalar();
 				return result > 0;
 			}
 			catch (SQLiteException) {
