+++ gestadh45.dao/VilleDao.cs
@@ -83,7 +83,7 @@ namespace gestadh45.dao
 			cmd.Parameters.Add(paramLibelle);
 
 			try {
-				var result = (int)cmd.ExecuteScalar();
+				var result = (long)cmd.ExecuteScalar();
 				return result > 0;
 			}
 			catch (SQLiteException) {
@@ -102,7 +102,7 @@ namespace gestadh45.dao
 			cmd.Parameters.Add(paramId);
 
 			try {
-				var result = (int)cmd.ExecuteScalar();
+				var result = (long)cmd.ExecuteScalar();
 				return result > 0;
 			}
 			catch (SQLiteException) {
